Refactor: 친구 추가 경계 정리 #153
